除了通知，其他都做好了

+++ src/main/java/com/bluemsun/entity/User.java
@@ -21,7 +21,7 @@ public class User implements Serializable
     String photo;
     Long phone;
     Integer admin;
-    Byte banned;
+    Integer banned;
     Long uuid;
     Timestamp createTime;
     Timestamp updateTime;
